+++ src/DDTrace/Log/LoggerInterface.php
@@ -8,18 +8,6 @@ namespace DDTrace\Log;
 interface LoggerInterface
 {
     /**
-     * Const list from https://www.php-fig.org/psr/psr-3/
-     */
-    const EMERGENCY = 'emergency';
-    const ALERT     = 'alert';
-    const CRITICAL  = 'critical';
-    const ERROR     = 'error';
-    const WARNING   = 'warning';
-    const NOTICE    = 'notice';
-    const INFO      = 'info';
-    const DEBUG     = 'debug';
-
-    /**
      * Logs a message at the debug level.
      *
      * @param string $message
@@ -48,4 +36,10 @@ interface LoggerInterface
      * @return void
      */
     public function error($message, array $context = array());
+
+    /**
+     * @param string $level
+     * @return bool
+     */
+    public function isLevelActive($level);
 }